Refactor e-invoicing to allow more types for quotes, po, ...
