+++ tests/Feature/EInvoice/FatturaPATest.php
@@ -11,7 +11,7 @@
 
 namespace Tests\Feature\EInvoice;
 
-use App\Services\Invoice\EInvoice\FatturaPA;
+use App\Services\EInvoicing\Standards\FatturaPA;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 use Illuminate\Routing\Middleware\ThrottleRequests;
 use Tests\MockAccountData;
@@ -42,7 +42,7 @@ class FatturaPATest extends TestCase
         $xml = $fat->run();
 
         // nlog($xml);
-        
+
         $this->assertnotNull($xml);
     }
 }
